Skip entire test suites of browser tests for Windows, Mac and Linux
Skipping individual tests is not reliable enough and, besides, the
huge list of tests running is killing the MacOS bot, so we'll try
skipping entire test suites first to see how it works.

And if this is not yet reliable enough to bring the MacOS bots down
to sane times, we'll skip all tests in that platform for now.